Rename useInput -> use-input
